# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Multi-Institution Admission and Student Onboarding Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateCurrentUser
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (super admin)
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: institutionId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/sync:
    post:
      operationId: syncUser
      tags: [users]
      summary: Sync user from auth provider (upsert on login)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSync"
      responses:
        "200":
          description: Synced user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── INSTITUTIONS ─────────────────────────────────────────────────────────────
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students
      parameters:
        - name: institutionId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create student record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Created student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update student record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}/guardian:
    get:
      operationId: getGuardian
      tags: [students]
      summary: Get student guardian info
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Guardian
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guardian"
    put:
      operationId: upsertGuardian
      tags: [students]
      summary: Create or update guardian
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuardianInput"
      responses:
        "200":
          description: Guardian
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guardian"
  /students/{id}/medical:
    put:
      operationId: upsertStudentMedical
      tags: [students]
      summary: Create or update medical screening data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicalInput"
      responses:
        "200":
          description: Medical record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medical"

  # ─── DOCUMENTS ────────────────────────────────────────────────────────────────
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      parameters:
        - name: applicationId
          in: query
          schema:
            type: ["integer", "null"]
        - name: institutionId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Initiate payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get payment details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}/verify:
    post:
      operationId: verifyPayment
      tags: [payments]
      summary: Manually verify payment (finance officer)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Verified payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /receipts/{id}:
    get:
      operationId: getReceipt
      tags: [payments]
      summary: Get receipt for payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receipt"

  # ─── ALLOCATIONS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, clerkId, email, role, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        fullName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [super_admin, institution_admin, admission_officer, finance_officer, storekeeper, parent]
        institutionId:
          type: ["integer", "null"]
        institution:
          $ref: "#/components/schemas/InstitutionSummary"
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        role:
          type: string
        institutionId:
          type: ["integer", "null"]
    UserSync:
      type: object
      required: [clerkId, email]
      properties:
        clerkId:
          type: string
        email:
          type: string
        fullName:
          type: ["string", "null"]

    # ─── INSTITUTIONS ─────────────────────────────────────────────────────────
...
    InstitutionSummary:
      type: object
      required: [id, name, level]
      properties:
        id:
          type: integer
        name:
          type: string
        level:
          type: string
        logoUrl:
          type: ["string", "null"]
...
    Student:
      type: object
      required: [id, fullName, gender, createdAt]
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        fullName:
          type: string
        gender:
          type: string
          enum: [male, female, other]
        dateOfBirth:
          type: ["string", "null"]
        assessmentNumber:
          type: ["string", "null"]
        admissionNumber:
          type: ["string", "null"]
        religion:
          type: ["string", "null"]
        county:
          type: ["string", "null"]
        homeAddress:
          type: ["string", "null"]
        previousSchool:
          type: ["string", "null"]
        passportPhotoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    StudentInput:
      type: object
      required: [applicationId, fullName, gender]
      properties:
        applicationId:
          type: integer
        fullName:
          type: string
        gender:
          type: string
          enum: [male, female, other]
        dateOfBirth:
          type: string
        assessmentNumber:
          type: string
        admissionNumber:
          type: string
        religion:
          type: string
        county:
          type: string
        homeAddress:
          type: string
        previousSchool:
          type: string
        passportPhotoUrl:
          type: string
    StudentUpdate:
      type: object
      properties:
        fullName:
          type: string
        gender:
          type: string
        dateOfBirth:
          type: string
        religion:
          type: string
        county:
          type: string
        homeAddress:
          type: string
        previousSchool:
          type: string
        passportPhotoUrl:
          type: string
        admissionNumber:
          type: string
    Guardian:
      type: object
      required: [id, studentId, parentName, relationship, phone]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        parentName:
          type: string
        relationship:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        occupation:
          type: ["string", "null"]
        residentialAddress:
          type: ["string", "null"]
        emergencyContact:
          type: ["string", "null"]
    GuardianInput:
      type: object
      required: [parentName, relationship, phone]
      properties:
        parentName:
          type: string
        relationship:
          type: string
        phone:
          type: string
        email:
          type: string
        occupation:
          type: string
        residentialAddress:
          type: string
        emergencyContact:
          type: string
    Medical:
      type: object
      required: [id, studentId]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        height:
          type: ["number", "null"]
        weight:
          type: ["number", "null"]
        vision:
          type: ["string", "null"]
        allergies:
          type: ["string", "null"]
        medicalConditions:
          type: ["string", "null"]
    MedicalInput:
      type: object
      properties:
        height:
          type: number
        weight:
          type: number
        vision:
          type: string
        allergies:
          type: string
        medicalConditions:
          type: string

    # ─── DOCUMENTS ────────────────────────────────────────────────────────────
...
    Payment:
      type: object
      required: [id, applicationId, amount, method, status, createdAt]
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        receiptNumber:
          type: ["string", "null"]
        transactionId:
          type: ["string", "null"]
        amount:
          type: number
        method:
          type: string
          enum: [mpesa, airtel_money, bank_transfer, card]
        status:
          type: string
          enum: [pending, confirmed, failed]
        breakdown:
          $ref: "#/components/schemas/PaymentBreakdown"
        createdAt:
          type: string
    PaymentBreakdown:
      type: object
      properties:
        tuitionFee:
          type: ["number", "null"]
        boardingFee:
          type: ["number", "null"]
        activityFee:
          type: ["number", "null"]
        admissionFee:
          type: ["number", "null"]
        itemsTotal:
          type: ["number", "null"]
        otherFees:
          type: ["number", "null"]
    PaymentInput:
      type: object
      required: [applicationId, amount, method]
      properties:
        applicationId:
          type: integer
        amount:
          type: number
        method:
          type: string
          enum: [mpesa, airtel_money, bank_transfer, card]
        transactionId:
          type: string
        breakdown:
          $ref: "#/components/schemas/PaymentBreakdown"
    Receipt:
      type: object
      required: [id, paymentId, receiptNumber, amount, issuedAt]
      properties:
        id:
          type: integer
        paymentId:
          type: integer
        receiptNumber:
          type: string
        studentName:
          type: string
        institutionName:
          type: string
        amount:
          type: number
        issuedAt:
          type: string

    # ─── ALLOCATIONS ──────────────────────────────────────────────────────────
...