# lib/api-zod/src/generated/types/paymentBreakdown.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Smart Multi-Institution Admission and Student Onboarding Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentBreakdown {
  /** @nullable */
  tuitionFee?: number | null;
  /** @nullable */
  boardingFee?: number | null;
  /** @nullable */
  activityFee?: number | null;
  /** @nullable */
  admissionFee?: number | null;
  /** @nullable */
  itemsTotal?: number | null;
  /** @nullable */
  otherFees?: number | null;
}